+++ Source/SACircularCountdown.xcodeproj/xcshareddata/xcschemes/SACircularCountdown.xcscheme
@@ -14,9 +14,9 @@
             buildForAnalyzing = "YES">
             <BuildableReference
                BuildableIdentifier = "primary"
-               BlueprintIdentifier = "F8F9A5341C2F039200EE7C24"
-               BuildableName = "SACircularCountdownDemo.app"
-               BlueprintName = "SACircularCountdownDemo"
+               BlueprintIdentifier = "F85CB7562043D09C008B9544"
+               BuildableName = "SACircularCountdown.framework"
+               BlueprintName = "SACircularCountdown"
                ReferencedContainer = "container:SACircularCountdown.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
@@ -29,26 +29,7 @@
       language = ""
       shouldUseLaunchSchemeArgsEnv = "YES">
       <Testables>
-         <TestableReference
-            skipped = "NO">
-            <BuildableReference
-               BuildableIdentifier = "primary"
-               BlueprintIdentifier = "F8F9A5481C2F039200EE7C24"
-               BuildableName = "SACircularCountdownTests.xctest"
-               BlueprintName = "SACircularCountdownTests"
-               ReferencedContainer = "container:SACircularCountdown.xcodeproj">
-            </BuildableReference>
-         </TestableReference>
       </Testables>
-      <MacroExpansion>
-         <BuildableReference
-            BuildableIdentifier = "primary"
-            BlueprintIdentifier = "F8F9A5341C2F039200EE7C24"
-            BuildableName = "SACircularCountdownDemo.app"
-            BlueprintName = "SACircularCountdownDemo"
-            ReferencedContainer = "container:SACircularCountdown.xcodeproj">
-         </BuildableReference>
-      </MacroExpansion>
       <AdditionalOptions>
       </AdditionalOptions>
    </TestAction>
@@ -63,16 +44,15 @@
       debugDocumentVersioning = "YES"
       debugServiceExtension = "internal"
       allowLocationSimulation = "YES">
-      <BuildableProductRunnable
-         runnableDebuggingMode = "0">
+      <MacroExpansion>
          <BuildableReference
             BuildableIdentifier = "primary"
-            BlueprintIdentifier = "F8F9A5341C2F039200EE7C24"
-            BuildableName = "SACircularCountdownDemo.app"
-            BlueprintName = "SACircularCountdownDemo"
+            BlueprintIdentifier = "F85CB7562043D09C008B9544"
+            BuildableName = "SACircularCountdown.framework"
+            BlueprintName = "SACircularCountdown"
             ReferencedContainer = "container:SACircularCountdown.xcodeproj">
          </BuildableReference>
-      </BuildableProductRunnable>
+      </MacroExpansion>
       <AdditionalOptions>
       </AdditionalOptions>
    </LaunchAction>
@@ -82,16 +62,15 @@
       savedToolIdentifier = ""
       useCustomWorkingDirectory = "NO"
       debugDocumentVersioning = "YES">
-      <BuildableProductRunnable
-         runnableDebuggingMode = "0">
+      <MacroExpansion>
          <BuildableReference
             BuildableIdentifier = "primary"
-            BlueprintIdentifier = "F8F9A5341C2F039200EE7C24"
-            BuildableName = "SACircularCountdownDemo.app"
-            BlueprintName = "SACircularCountdownDemo"
+            BlueprintIdentifier = "F85CB7562043D09C008B9544"
+            BuildableName = "SACircularCountdown.framework"
+            BlueprintName = "SACircularCountdown"
             ReferencedContainer = "container:SACircularCountdown.xcodeproj">
          </BuildableReference>
-      </BuildableProductRunnable>
+      </MacroExpansion>
    </ProfileAction>
    <AnalyzeAction
       buildConfiguration = "Debug">
